+++ apps/backend/src/modules/reviews/review.model.ts
@@ -1,7 +1,12 @@
+import { Model, type RelationMappings } from "objection";
+
 import {
 	AbstractModel,
 	DatabaseTableName,
 } from "~/libs/modules/database/database.js";
+
+import { UserModel } from "../users/user.model.js";
+import { type UserAuthResponseDto } from "../users/users.js";
 
 class ReviewModel extends AbstractModel {
 	public static override get tableName(): string {
@@ -16,7 +21,24 @@ class ReviewModel extends AbstractModel {
 
 	public routeId!: null | number;
 
+	public user!: Pick<UserAuthResponseDto, "firstName" | "id" | "lastName"> & {
+		avatar: null | { url: null | string };
+	};
+
 	public userId!: number;
+
+	public static relationMappings(): RelationMappings {
+		return {
+			user: {
+				join: {
+					from: `${DatabaseTableName.REVIEWS}.user_id`,
+					to: `${DatabaseTableName.USERS}.id`,
+				},
+				modelClass: UserModel,
+				relation: Model.BelongsToOneRelation,
+			},
+		};
+	}
 }
 
 export { ReviewModel };