+++ apps/frontend/src/modules/routes/slices/route-details.slice.ts
@@ -2,19 +2,33 @@ import { createSlice } from "@reduxjs/toolkit";
 
 import { DataStatus } from "~/libs/enums/enums.js";
 import { type ValueOf } from "~/libs/types/types.js";
+import { type ReviewGetByIdResponseDto } from "~/modules/reviews/reviews.js";
 
 import { type RouteGetByIdResponseDto } from "../libs/types/types.js";
-import { deleteImage, getById, patch, uploadImage } from "./actions.js";
+import {
+	createReview,
+	deleteImage,
+	getById,
+	getReviews,
+	patch,
+	uploadImage,
+} from "./actions.js";
 
 type State = {
 	dataStatus: ValueOf<typeof DataStatus>;
 	patchStatus: ValueOf<typeof DataStatus>;
+	reviewCreateStatus: ValueOf<typeof DataStatus>;
+	reviews: ReviewGetByIdResponseDto[];
+	reviewsDataStatus: ValueOf<typeof DataStatus>;
 	route: null | RouteGetByIdResponseDto;
 };
 
 const initialState: State = {
 	dataStatus: DataStatus.IDLE,
 	patchStatus: DataStatus.IDLE,
+	reviewCreateStatus: DataStatus.IDLE,
+	reviews: [],
+	reviewsDataStatus: DataStatus.IDLE,
 	route: null,
 };
 
@@ -40,6 +54,28 @@ const { actions, name, reducer } = createSlice({
 		});
 		builder.addCase(patch.rejected, (state) => {
 			state.patchStatus = DataStatus.REJECTED;
+		});
+
+		builder.addCase(getReviews.pending, (state) => {
+			state.reviewsDataStatus = DataStatus.PENDING;
+		});
+		builder.addCase(getReviews.fulfilled, (state, { payload }) => {
+			state.reviews = payload.data;
+			state.reviewsDataStatus = DataStatus.FULFILLED;
+		});
+		builder.addCase(getReviews.rejected, (state) => {
+			state.reviewsDataStatus = DataStatus.REJECTED;
+		});
+
+		builder.addCase(createReview.pending, (state) => {
+			state.reviewCreateStatus = DataStatus.PENDING;
+		});
+		builder.addCase(createReview.fulfilled, (state, { payload }) => {
+			state.reviews.unshift(payload.data);
+			state.reviewCreateStatus = DataStatus.FULFILLED;
+		});
+		builder.addCase(createReview.rejected, (state) => {
+			state.reviewCreateStatus = DataStatus.REJECTED;
 		});
 
 		builder.addCase(deleteImage.fulfilled, (state, action) => {